API

Base.read — Method

read(file::AbstractString, ::AbstractPlateReader; channels)

Read data from plate reader file file of assuming the format of AbstractPlateReader.

Arguments:

file::AbstractString: File to read from.

::AbstractPlateReader: Plate reader type.

channels: Vector of channels (as strings) to be read in. Defaults to nothing (all channels).

Returns:

Dict{String, DataFrame}: A dictionary with keys being the channel names and values being the dataframes containing the data for each channel.

String: The plate reader raw metadata.

Base.summary — Method

summary(file, ptype::AbstractESMDataType; plot=false)

Summarise a data file (.esm, plate reader, .fcs, etc.).

Arguments

file::AbstractString: The data file to be summarised.

ptype::AbstractESMDataType: The type of data file.

plot::Bool=false: Produce plots of the data. Defaults to false.

DataFrames.transform — Method

transform(data::DataFrame, method::AbstractTransformMethod)

Apply the specified transformation to data.

Arguments:

data::DataFrame: A DataFrame of Flow cytometry data.

method::AbstractTransformMethod: The method and settings to use for transformation.

ESM.Arcsinh — Method

Arcsinh()

Transforms data to inverse hyperbolic sine scale.

ESM.Bound — Method

Bound(; min = -Inf, max = Inf)

Clamps data outside of the interval [min, max].

When untransforming, values that were clamped to `min` are mapped to `-Inf` and values clamped to `max` are mapped to `Inf`.

ESM.Hyperlog — Method

Hyperlog(; W = 1.0, T = 262144, M = 4.5, A = 0)

Transforms data to hyperlog scale. The hyperlog transformation is a continuous, monotonic transformation that behaves like a logarithm for large values and like a linear function for small values. It is defined by the parameters W, T, M, and A.

The hyperlog function is defined by inverting the function:

f(x) = a * exp(b * x) + c * x - f

where the parameters a, b, c, and f are computed from W, T, M, and A.

Arguments:

W: The number of linear decades. Must be greater than 0 and less than or equal to M/2.

T: The top of the scale. Must be greater than 0.

M: The number of decades the logarithmic scale covers. Must be greater than 0.

A: The additional negative decades. Must be greater than or equal to -W and less than or equal to M - 2*W.

ESM.Linear — Method

Linear(A, T)

Transforms data to linear scale from the interval [-A, T] to [0, 1].

ESM.Log — Method

Log()

Transforms data to natural (base e) logarithm scale.

ESM.Log10 — Method

Log10()

Transforms data to base 10 logarithm scale.

ESM.Log1p — Method

Log1p()

Transforms data to natural logarithm of (1 + x).

ESM.Log2 — Method

Log2()

Transforms data to base 2 logarithm scale.

ESM.Logicle — Method

Logicle(; W = 1.0, T = 262144, M = 4.5, A = 0)

The Logicle transformation is a continuous, monotonic transformation that behaves like a logarithm for large values and like a linear function for small values. It is defined by the parameters W, T, M, and A.

It is defined by inverting the function:

f(x) = a * exp(b * x) + c * exp(-d * x) + f

where the parameters a, b, c, d, and f are computed from W, T, M, and A.

Arguments:

W: The number of linear decades. Must be greater than or equal to 0

T: The total range of the transformation. Must be greater than 0

M: The number of decades to be compressed. Must be greater than 0

A: The additional negative decades. Must be greater than or equal to -W and less than or equal to M - 2*W.

ESM.at — Method

at(df, range_df, target_value)

Return the values of df at the index of range_df where range_df is equal to target_value.

If no value in range_df is equal to target_value, return the last recording before the target_value.

If range_df has one column, the same index will be applied to all columns of df. If range_df has multiple columns, the column names should match those of df and a new index will be computed for each column.

Arguments:

df::DataFrame: The DataFrame you are getting the indexes from.

range_df: The data frame by which to filter.

target_value::Number: Defines the threshold value for filtering.

ESM.at_time — Method

at_time(df, time_col, time_point)

Returns values at a specific time point. If no value at specific timepoint, return the last recording before the timepoint. If time_point < minimum(time_col), return nothing.

Arguments:

df::DataFrame: DataFrame on which to work.

time_col::DataFrame: Time values in a DataFrame.

time_point::Float64: Time point in mins at which to report the measurement.

ESM.between — Method

between(df, range_df; min_value=-Inf, max_value=Inf)
between(df; min_value=-Inf, max_value=Inf)

Replace out of range values with missing.

Out of range values are determined by min_value .<= range_col .<= max_value. If range_col is not provided, by min_value .<= df .<= max_value.

If range_col has one column, the associate row indexes with min_value .<= range_col .<= max_value are broadcast along all columns of df. If range_col has multiple columns, the column names should match those of df.

Arguments:

df::DataFrame: DataFrame to subset.

range_df::DataFrame: DataFrame of values to filter by.

min_value=-Inf: Minimum value.

max_value=Inf: Maximum value.

ESM.between_times — Method

between_times(df, time_col; mint=-Inf, maxt=Inf)

Returns the DataFrame between two timepoints.

All other data is set to missing.

Arguments:

df::DataFrame: DataFrame to work on.

time_col::DataFrame: DataFrame of time values.

mint::Float64=-Inf: Minimum time in mins.

maxt::Float64=Inf: Max time in mins.

ESM.calibrate — Method

calibrate(data, time_col, method::AbstractCalibrationMethod)

Calibrate data, for example, to remove background OD signal.

Arguments:

data::DataFrame: DataFrame containing the data to be calibrated.

time_col::Union{DataFrame, Nothing}: Time column associated with the data.

method::AbstractCalibrationMethod: Method to use for calibration.

Keywords:

offset::Number: Value to add to the calibrated data after calibration (default: 0).

ESM.comonicon_install — Method

comonicon_install(;kwargs...)

Install the CLI manually. This will use the default configuration in Comonicon.toml, if it exists. For more detailed reference, please refer to Comonicon documentation.

ESM.comonicon_install_path — Method

comonicon_install_path(;[yes=false])

Install the PATH and FPATH to your shell configuration file. You can use comonicon_install_path(;yes=true) to skip interactive prompt. For more detailed reference, please refer to Comonicon documentation.

ESM.doubling_time — Method

doubling_time(df, time_col, method::AbstractGrowthRateMethod; kwargs...)

Calculates the doubling time of a given DataFrame. Returns result in minutes.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod: Method to use for calculating growth rate.

kwargs...: Additional keyword arguments to pass to the growth rate method.

ESM.event_count — Method

event_count(data)

Count the number of events in the flow cytometry data.

Arguments:

data::Dict: Dict returned by to_rfi.

ESM.expand_group — Method

expand_group(group::AbstractString)

Expand a condensed group name into the sample IDs it contains. For example, if the group is "plate0[1,2][a:c]2", it will return all sample IDs from plate 1 to 2 and wells a2, b2, and c2.

Arguments:

group::AbstractString: The group name to expand.

Returns:

Vector{String}: A vector of sample IDs contained in the group.

ESM.expand_groups — Method

expand_groups(groups::AbstractString)

Iteratively call expand_group on a comma-separated list of group names.

ESM.extract_flow — Method

extract_flow(fcs,chan)

Extract the metadata from a FCSFiles.FlowSample.

Arguments:

fcs: The parsed FCS file.

chan::String: The channel being read from.

ESM.filter_col — Method

filter_col(df, reg_l)

Filters a df by a set of columns or parts of columns given in regl. regl is joined to form regex using |.

Arguments:

df::DataFrame: DataFrame to filter.

reg_l::Vector{String}: Vector of strings to filter by.

ESM.filter_row — Method

filter_row(es,group)

Filter the original es.samples dataframe by a specific group defined in the esm.

Arguments:

es::esm_zones: esm zones data type.

group::Union{String,Vector}: what group/groups to filter by.

ESM.find_group — Method

find_group(es,grn)

Finds a specific group from the original es.groups dataframe. Returns the sample names.

Arguments:

es::esm_zones: The data set to search.

grn::String: A string of a group name which can be used to filter the original dataframe.

ESM.fit_ellipse — Method

fit_ellipse(x, y) → a, b, θ, x₀, y₀, p

Fit an ellipse into the 2D data defined by (x, y) using ordinary least squares.

Return: semi-major axis length, semi-minor axis length, ellipse rotation, center coordinates and a parameter container for quadratic form of ellipse. The first five parameters represent the parametric form of an ellipse, which may have an arbitrary rotation and translation w.r.t. the origin:

x(t) = cos(θ)*a*cos(t) - sin(θ)*b*sin(t) + x₀
y(t) = sin(θ)*a*cos(t) + cos(θ)*b*sin(t) + y₀
t ∈ [0, 2π)

Use ellipse_from_parametric(a, b, θ, x₀, y₀) → x, y to construct points along the ellipse. Notice that always two possible ellipses can fit the data, one with a, b, θ and one with b, a, θ-π/2.

The quadratic form of the ellipse is created using p, and is given by the form (x^2, x*y, y^2, x, y) ⋅ p = 1.

Code modified from: https://www.matecdev.com/posts/julia-least-squares-qr.html using a lot of stuff from: https://en.wikipedia.org/wiki/Ellipse#General_ellipse

ESM.fluorescence — Function

fluorescence(data_fl, time_fl, data_od, time_od, method::AbstractFluorescenceMethod)

Calculate fluorescence per cell (normalised by OD).

Arguments

data_fl: a DataFrame of fluorescence measurements.

time_fl: a DataFrame of times for the fluorescence measurements.

data_od: a DataFrame of OD measurements.

time_od: a DataFrame of times for the OD measurements.

method: the method to use for calculating fluorescence per cell.

If method is RatioAtMaxGrowth, any keyword arguments are passed to growth_rate.

ESM.form_df — Method

form_df(df)

Creates the data frame from the dataframe it is passed. This is to turn the es.samples.values arrays into dfs.

Arguments:

df::DataFrame: A dataframe of es.samples to parse.

ESM.gate — Function

gate(data, method::AbstractGatingMethod)

Filter data to only include events within the gate defined by method.

Arguments:

data::Dict: Dict returned by to_rfi.

method::AbstractGatingMethod: The method and settings to use for gating.

ESM.gated_proportion — Method

gated_proportion(data, gate)
gated_proportion(data_before, data_after)

Calculate the proportion of events remaining after gating.

Arguments:

data::Dict: Dict returned by to_rfi.

gate::AbstractGatingMethod: A gating method to report on.

data_before::Dict: Dict returned by to_rfi before gating.

data_after::Dict: Dict returned by to_rfi after gating.

ESM.get_metadata — Method

get_metadata()

Generate the metadata for a new ESM file, such as the version of ESM used to create it.

ESM.growth_rate — Method

growth_rate(df, time_col, method::AbstractGrowthRateMethod)

Calculates the growth rate of a given dataframe. Returns in min^-1 using base e.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod: Method to use for calculating growth rate.

Keywords:

recalibrate: Whether to recalibrate the data using calibrate before calculating growth rate. Default is :negative (only if negative values are present in the well). Available options are :negative, true, and false.

offset: If data is recalibrated, this is the offset applied after calibration. Default is 0.001.

plot_directory: If provided, this is the directory to save plots of the growth curves with the fitted growth rate. Default is nothing (no plots saved). If :temp, plots will be saved to a temporary directory.

ESM.lag_time — Method

lag_time(df, time_col, method::AbstractGrowthRateMethod)

Calculates the lag time of the growth curve. This is given as the t-axis intercept of the tangent at the point of maximum growth (defined by the growth rate, time to max growth, and od at max growth).

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod.

ESM.max_od — Method

max_od(df, time_col, method::AbstractGrowthRateMethod)

Calculates the maximum OD of a given dataframe. This may depend on the growth rate method used.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod.

ESM.od_at_max_growth — Method

od_at_max_growth(df, time_col, method::AbstractGrowthRateMethod)

The OD at which growth is maximised.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod.

ESM.produce_views — Method

produce_views(es, trans_meta_map;to_out=[])

Produces the subset of views specified in to_out or all views if unspecified.

Arguments:

es::esm_zones: The esm_zones object

trans_meta_map::Dict: A dictionary mapping transformations to their names.

to_out::Vector{String}: A list of all the views to be produced.

ESM.read_data — Method

read_data(file::AbstractString)

Read the data from path file into the correct structure.

ESM.read_esm — Method

read_esm(file::AbstractString)

Parse an esm file found at file into an esm_zones object.

ESM.read_flow — Method

read_flow(samples, sample_dict,channels,broad_g,channel_map)

Read the flow cytometry data from a file.

Arguments:

samples::DataFrame: DataFrame of samples to read from.

sample_dict::Dict: Output sample Dict.

channels::Vector: Channels to read.

broad_g::Vector: The physical group that this is part of.

channel_map::Dict: The names to replace the old channel names with.

ESM.read_multipr_file — Method

read_multipr_file(file,ptype,channels,channel_map)

Function for reading files containing multiple reads from multple channels from a single CSV file. Returns a Dictionary of DataFrames with keys being the channels.

Arguments:

file::String: File to read from.

ptype::String: The plate reader type being used.

channels::Vector: Channels to be read in.

channel_map: The new names for the channels.

ESM.read_pr — Method

read_pr(samples,sample_dict,channels,broad_g,channel_map)

Reads plate reader data.

Arguments:

samples::DataFrame: Dataframe of samples to be read into sample_dict.

sample_dict::Dict: A dictionary of the samples.

channels::Vector: Channels to use.

broad_g::Vector: The larger plate that this belongs to.

channel_map::Dict: What to rename the channel to.

ESM.remove_subcols — Method

remove_subcols(df, sub)

Gets rid of certain sub col ids in the col names.

e.g.

remove flo
a1.flo -> a1

ESM.sexp_to_nested_list — Method

sexp_to_nested_list(sexp,es,trans_meta_map)

Recursively converts the parsed equations to julia code and produces the correct dataframes over which to operate. This also calls any unprocessed transformations.

Arguments:

sexp: Expression or part of expression to be decomposed.

es: The esm_zones data type that contains the data.

transmetamap: The transformation map of the parsed transformations.

ESM.smooth — Function

smooth(data, time_col, method::AbstractSmoothingMethod)

Smooth that data using the specified method.

Arguments

data: A DataFrame of measurements.

time_col: The column of time data corresponding to the measurements in data.

method: The method to use for smoothing.

ESM.summarise — Function

esm summarise

Summarise a data file (.esm, plate reader, .fcs, etc.).

Args

file: The data file to be summarised.

Options

-t, --type=<String>: The type of data file. Options are "auto" (default), "esm", "spectramax", "biotek", "tecan", "bmg", "generic", "fcs". If "auto" is selected, the type will be inferred from the file extension (or raise an error if not possible).

Flags

-p, --plot: Produce plots of the data. Not available for --type=esm.

-c, --csv: Save the data as CSV files. Not available for --type=esm.

ESM.template — Function

esm template

Produce a template excel file for data entry into the ESM.

Options

-o, --output-path=<String>: The path to create the template in. Defaults to ESM.xlsx in the current directory.

ESM.time_to_max_growth — Method

time_to_max_growth(df, time_col, method::AbstractGrowthRateMethod)

Calculates the time at which growth is maximised.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod.

ESM.to_rfi — Method

to_rfi(sample_name;chans=[])

Calculates relative fluorescence of given sample.

Arguments:

sample_name::String: channel to use.

chans::Vector: vector of channels to keep.

ESM.translate — Function

esm translate

Translates the completed .xlsx template file to a .esm file.

Args

input: The completed .xlsx template file to be read.

output: The filepath/destination for the .esm file.

ESM.untransform — Method

untransform(data::DataFrame, method::AbstractTransformMethod)

Reverse a previously applied transformation on data.

Arguments:

data::DataFrame: A DataFrame of Flow cytometry data.

method::AbstractTransformMethod: The method and settings to use for transformation.

ESM.view_to_csv — Method

view_to_csv(es,trans_meta_map;out_dir,to_out)

Produces views and writes them to CSV files.

Arguments:

es::esm_zones: The original esm struct

trans_meta_map::Dict: The transformations that have been parsed at the top level.

outdir::String: The specified output dir - defaults to nothing.

to_out::Vector{String}: The views to be processed.

ESM.views — Function

esm views

Produce and save the views from a .esm file.

Args

esm_file: The .esm file to be read.

Options

-v, --view=<String>: The view to be produced. All views if not specified.

-o, --output-dir=<String>: The directory to save the output(s) to. Defaults to the current directory.

ESM.write_esm — Method

write_esm(data, file::AbstractString)

Write the esm data to the path file.