Command Line Interface

The command line interface allows you to interact with, create and edit .esm files. The functions are documented here and can also be seen by using esm -h. The help for each function can be seen by using esm summarise -h or esm template -h for example.

summarise

A typcial working begins with esm summarise, which lets you view your raw data. This can be helpful for decisions such as detecting contamination in blank wells, gating, etc.

ESM.summarise — Function

esm summarise

Summarise a data file (.esm, plate reader, .fcs, etc.).

Options

-f, --file=<String>: The data file to be summarised.

-t, --type=<String>: The type of data file. Options are "auto" (default), "esm", "spectramax", "biotek", "fcs". If "auto" is selected, the type will be inferred from the file extension (or raise an error if not possible).

Flags

-p, --plot: Produce plots of the data. Only available for some types.

template

To convert and collect your data into a .esm file, you first want to create an Excel file that describes your data and how it should be imported, processed and viewed. A template Excel file can be loaded using the esm template ... function. To see how the Excel file should be filled out, check out Getting Started with ESM or Excel Interface.

ESM.template — Function

esm template

Produce a template excel file for data entry into the ESM.

Options

-o, --output-path=<String>: The path to create the template in. Defaults to ESM.xlsx in the current directory.

translate

Once the template has been filled out and completed, it can be translated into a .esm file using the esm translate ... function.

ESM.translate — Function

esm translate

Translates the completed .xlsx template file to a .esm file.

Options

-e, --excel=<String>: The .xlsx template file to be read.

-t, --target=<String>: The name of the output .esm file.

views

To create the views from a .esm file, you can use the esm views ... function. This saves the views as .csv files or plots relevant figures.

ESM.views — Function

esm views

Produce and save the views from a .esm file.

Options

-e, --esm-file=<String>: The .esm file to be read.

-v, --view=<String>: The view to be produced. If not specified, all views will be produced.

-o, --output-dir=<String>: The directory to save the output(s) to. Defaults to the current directory.