Transforms

Flow cytometry data typically varies across multiple orders of magnitude and includes both positive and negative values. As such, you may want to apply any of a long list of different transformations. We have a selection of them implementated in ESM. They can be accessed through the transform(data, Transform()) and untransform(data, Transform()) functions.

DataFrames.transform — Function

transform(data::DataFrame, method::AbstractTransformMethod)

Apply the specified transformation to data.

Arguments:

data::DataFrame: A DataFrame of Flow cytometry data.

method::AbstractTransformMethod: The method and settings to use for transformation.

ESM.untransform — Function

untransform(data::DataFrame, method::AbstractTransformMethod)

Reverse a previously applied transformation on data.

Arguments:

data::DataFrame: A DataFrame of Flow cytometry data.

method::AbstractTransformMethod: The method and settings to use for transformation.

Methods

The different methods that can be used with the transform and untransform functions are outlined below.

ESM.Log — Function

Log()

Transforms data to natural (base e) logarithm scale.

ESM.Log10 — Function

Log10()

Transforms data to base 10 logarithm scale.

ESM.Log2 — Function

Log2()

Transforms data to base 2 logarithm scale.

ESM.Log1p — Function

Log1p()

Transforms data to natural logarithm of (1 + x).

ESM.Arcsinh — Function

Arcsinh()

Transforms data to inverse hyperbolic sine scale.

ESM.Linear — Function

Linear(A, T)

Transforms data to linear scale from the interval [-A, T] to [0, 1].

ESM.Bound — Function

Bound(; min = -Inf, max = Inf)

Clamps data outside of the interval [min, max].

When untransforming, values that were clamped to `min` are mapped to `-Inf` and values clamped to `max` are mapped to `Inf`.

ESM.Hyperlog — Function

Hyperlog(; W = 1.0, T = 262144, M = 4.5, A = 0)

Transforms data to hyperlog scale. The hyperlog transformation is a continuous, monotonic transformation that behaves like a logarithm for large values and like a linear function for small values. It is defined by the parameters W, T, M, and A.

The hyperlog function is defined by inverting the function:

f(x) = a * exp(b * x) + c * x - f

where the parameters a, b, c, and f are computed from W, T, M, and A.

Arguments:

W: The number of linear decades. Must be greater than 0 and less than or equal to M/2.

T: The top of the scale. Must be greater than 0.

M: The number of decades the logarithmic scale covers. Must be greater than 0.

A: The additional negative decades. Must be greater than or equal to -W and less than or equal to M - 2*W.

ESM.Logicle — Function

Logicle(; W = 1.0, T = 262144, M = 4.5, A = 0)

The Logicle transformation is a continuous, monotonic transformation that behaves like a logarithm for large values and like a linear function for small values. It is defined by the parameters W, T, M, and A.

It is defined by inverting the function:

f(x) = a * exp(b * x) + c * exp(-d * x) + f

where the parameters a, b, c, d, and f are computed from W, T, M, and A.

Arguments:

W: The number of linear decades. Must be greater than or equal to 0

T: The total range of the transformation. Must be greater than 0

M: The number of decades to be compressed. Must be greater than 0

A: The additional negative decades. Must be greater than or equal to -W and less than or equal to M - 2*W.

Implementation Details

If you want to implement a new transform method to be included in ESM, you need to:

Todo

This isn't right for how the transform method works, theres only one transform type

Open a pull request with the following code changes

Define a new function which returns a Transform object in src/flow/transform.jl

Document that method in the transformation documentation (this page)

If you are unsure how to do any of these steps, feel free to open an issue on GitHub asking for a new transformation and explaining how the transform should work.