Calibration

There are a few different methods for calibrating plate reader OD and fluorescence data. Each method uses the calibrate(data, blanks, Method()) function signature (calibrate(data, Method()) is also available if blanks aren't needed). These can be used either in the Julia ESM package diectly, or in the transformations in the Excel template.

ESM.calibrate — Function

calibrate(data, method::AbstractCalibrationMethod)
calibrate(data, blanks, method::AbstractCalibrationMethod)

Calibrate data, for example, to remove background OD signal.

Arguments:

data::DataFrame: DataFrame containing the data to be calibrated.

blanks::DataFrame: DataFrame containing the blank measurements.

method::AbstractCalibrationMethod: Method to use for calibration.

TimeseriesBlank

todo

It should probably interpolate if times are different

The TimeseriesBlank method averages across the blank wells, to get a single blank timeseries (assuming all the blanks were measured at the same timepoints). It then subtracts this from the data, timepoint by timepoint. So the first timepoint is calibrated by the average of all the blank wells at the first timepoint, the second is calibrated at the second timepoint and so on.

It can be called using calibrate(data, blanks, TimeseriesBlank()).

MeanBlank

The MeanBlank method averages both over wells and over time. This gives a single value to calibrate the data against which is subtracted from the data.

It can be called using calibrate(data, blanks, MeanBlank()).

MinBlank

The MinBlank method calculates a single calibration value as the minimum of all wells across all timepoints. It then subtracts this value form the data.

It can be called using calibrate(data, blanks, MinBlank()).

MinData

The MinData method doesn't require blanks. It calculates a single calibration value for each well, as the minimum of each well. It then subtracts these values from each well, so that the data for each well contains one zero (where the minimum was previously), and no negative points.

It can be called using calibrate(data, blanks, MinData()).

StartData

The StartData method doesn't require blanks. It calculates a single calibration value for each well, as the initial value of each well. It then subtracts these values from each well, so that the data begins at zero for all wells.

It can be called using calibrate(data, blanks, StartData()).

Implementation Details

If you want to implement a new calibration method to be included in ESM, you need to:

Open a pull request with the following code changes

Define a new struct for your method type in src/methods.jl

The type of that struct is a subtype of AbstractCalibrationMethod

Define a new method dispatch calibrate(data, blanks, ::NameOfNewMethodType)

Document that method in the calibration documentation (this page)

If you are unsure how to do any of these steps, feel free to open an issue on GitHub asking for a new calibration method and explaining how the method should work.