Growth Rate

There are a variety of methods for calculating growth rates (or doubling times). Each method uses the growth_rate(data, time_col, Method()) function signature (or doubling_time(data, time_col, Method())). These can be used in the transformations in the Excel template.

ESM.growth_rate — Function

growth_rate(df, time_col, method::AbstractGrowthRateMethod)

Calculates the growth rate of a given dataframe. Returns in min^-1 using base e.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod: Method to use for calculating growth rate.

Keywords:

recalibrate: Whether to recalibrate the data using calibrate before calculating growth rate. Default is :negative (only if negative values are present in the well). Available options are :negative, true, and false.

offset: If data is recalibrated, this is the offset applied after calibration. Default is 0.001.

ESM.doubling_time — Function

doubling_time(df, time_col, method::AbstractGrowthRateMethod; kwargs...)

Calculates the doubling time of a given DataFrame. Returns result in minutes.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod: Method to use for calculating growth rate.

kwargs...: Additional keyword arguments to pass to the growth rate method.

Other functions

Not only are methods available for calculating growth_rate and doubling_time, but there are also methods for calculating related summary statistics of time and OD at maximum growth, lagtime, and maximum OD. All these use the same function signature and their definitions depend on the specific growth rate method that is being used.

ESM.max_od — Function

max_od(df, time_col, method::AbstractGrowthRateMethod)

Calculates the maximum OD of a given dataframe. This may depend on the growth rate method used.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod.

ESM.time_to_max_growth — Function

time_to_max_growth(df, time_col, method::AbstractGrowthRateMethod)

Calculates the time at which growth is maximised.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod.

ESM.od_at_max_growth — Function

od_at_max_growth(df, time_col, method::AbstractGrowthRateMethod)

The OD at which growth is maximised.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod.

ESM.lag_time — Function

lag_time(df, time_col, method::AbstractGrowthRateMethod)

Calculates the lag time of the growth curve. This is given as the t-axis intercept of the tangent at the point of maximum growth (defined by the growth rate, time to max growth, and od at max growth).

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod.

Lagtime

Lagtime is always calculated using growth_rate, time_to_max_growth, and od_at_max_growth according to Zwietering et al. 1990. This method defines the lagtime as the x intercept of the tangent to the growth curve at maximum growth on a plot of log(OD/OD_0) where OD_0 is the first OD value in the data set. It is calculated as: time_to_max_growth - (1 / growth_rate) * ln(od_at_max_growth / OD_0). For the parameteric and regularization methods, OD_0 is determined from the fitted curve.

Endpoints

The Endpoints method is the same as fitting an exponential curve $y=A\exp(bt)$ to two points (start_time and end_time) and returning the growth rate $b$.

\[b = \frac{\ln (od(t_\text{end})) - \ln (od(t_\text{start}))}{t_\text{end} - t_\text{start}}\]

It can be called using growth_rate(data, time_col, Endpoints(start_time, end_time)) or doubling_time(data, time_col, Endpoints(start_time, end_time)). start_time and end_time are the start and end times of the exponential phase.

max_od - return the maximum OD in the data

time_to_max_growth - return the average of the first time before start_time and the first time before end_time

od_at_max_growth - returns the geometric mean of the first OD before start_time and the first OD before end_time

LinearOnLog

The LinearOnLog method log-transforms the data (removing any points ≤ 0), and then fits a straight line on all the data between start_time and end_time, returning the gradient of the line of best fit.

It can be called using growth_rate(data, time_col, LinearOnLog(start_time, end_time)) or doubling_time(data, time_col, LinearOnLog(start_time, end_time)). start_time and end_time are the start and end times of the exponential phase.

max_od - return the maximum OD in the data

time_to_max_growth - return the average of the first time before start_time and the first time before end_time

od_at_max_growth - returns the geometric mean of the first OD before start_time and the first OD before end_time

MovingWindow

The MovingWindow method allows you to use any of the above methods (Endpoints, or LinearOnLog) without defining the start and end points of the exponential phase. Instead, you can provide a number of timepoints (a.k.a. window_size, defaults to 10) and it will calculate the growth rate on all consequetive runs of that length and return the maximum growth rate.

By default, the growth rates on each window are calculated using the Endpoints method. This can be changed by supplying a method keyword argument to MovingWindow. The available options are :Endpoints (default), and :LinearOnLog.

It can be called using growth_rate(data, time_col, MovingWindow(window_size, method)) or doubling_time(data, time_col, MovingWindow(window_size, method)).

The other functions (max_od, time_to_max_growth, od_at_max_growth) are defined based on the choice of method and evaluated across the window which gives the maximum growth rate.

FiniteDiff

The FiniteDiff method log-transforms the data (removing any points ≤ 0) and then calculates the local gradients. It does this by traveling along the timepoints one-by-one, either calculating the central or one-sided finite difference. It then returns the maximum gradient.

\[\frac{\ln (od(t_{i})) - \ln (od(t_{i-1}))}{t_{i} - t_{i-1}} \quad \text{one-sided} \\ \frac{\ln (od(t_{i+1})) - \ln (od(t_{i-1}))}{t_{i+1} - t_{i-1}} \quad \text{central}\]

It can be called using growth_rate(data, time_col, FiniteDiff()) (defaults to central) or any of growth_rate(data, time_col, FiniteDiff(method=:onesided)), or growth_rate(data, time_col, FiniteDiff(method=:central)). You can also call doubling_time with either of the FiniteDiff methods.

max_od - return the maximum OD in the data

time_to_max_growth - return the time at the centre of the finite differencing interval when growth is maximised.

od_at_max_growth - return the geometric mean of the OD at either end of the finite differencing interval when growth is maximised.

Parameteric Models

You also can fit a range of parametric models to calculate growth rates in ESM. All fits are done on data after a ${y=ln(OD/OD_0)}$ transformation (with negative OD values removed).

growth_rate - return the parameter $\mu$

lagtime - return the parameter $\lambda$

max_od - return the upper limit $exp(A)$ of the growth curve ($+OD_0$)

time_to_max_growth - return the fitted time at maximum growth

od_at_max_growth - return the fitted OD at maximum growth

Logistic

The Logistic method fits the following curve.

\[\frac{A}{1 + \exp(\frac{4\mu}{A} (\lambda - t) + 2)}\]

This form is derived from Zwietering et al. 1990[1].

It can be called using growth_rate(data, time_col, Logistic()) or doubling_time(data, time_col, Logistic()).

Gompertz

The Gompertz method fits the following curve.

\[A \exp(-\exp(\frac{\mu e}{A} (\lambda - t) + 1))\]

This form is derived from Zwietering et al. 1990[1].

It can be called using growth_rate(data, time_col, Gompertz()) or doubling_time(data, time_col, Gompertz()).

Modified Gompertz

The ModifiedGompertz method fits the following curve.

\[A \exp(-\exp(\frac{\mu e}{A} (\lambda - t) + 1)) + A \exp(\nu (t-t_\text{shift}))\]

This form is derived from Kahm et al. 2010[2].

It can be called using growth_rate(data, time_col, ModifiedGompertz()) or doubling_time(data, time_col, ModifiedGompertz()).

Richards

The Richards method fits the following curve.

\[A \left(\left(1 + \exp(\nu) \exp(1 + \exp(\nu)) \exp\left(\frac{\mu}{A} (1 + \exp(\nu))^{1 + \exp(-\nu)} (\lambda - t)\right)\right) ^ {-\exp(-\nu)}\right)\]

This form is derived from Zwietering et al. 1990[1]. The original form uses $\nu\in(0,\inf)$. We have transformed to use $\exp(\nu)$ instead so that no constraints need to be applied to $\nu$.

It can be called using growth_rate(data, time_col, Richards()) or doubling_time(data, time_col, Richards()).

Regularization

For the Regularization method, the data is log scaled (negative points removed) and smoothed using regularization, before being interpolated by a cubic spline. The point where the derivative of this smooth cubic spline is maximised determines the growth rate.

This method uses the RegularizationSmooth() method of DataInterpolations.jl, see here.

It can be called using growth_rate(data, time_col, Regularization()) or doubling_time(data, time_col, Regularization()).

max_od - returns the maximum value of the regularization within the time interval

time_to_max_growth - return the time where the derivative of the regularization is maximised

od_at_max_growth - return the regularized OD at time_to_max_growth

Implementation Details

If you want to implement a new growth rate method to be included in ESM, you need to:

Open a pull request with the following code changes

Define a new struct for your method type in src/methods.jl

The type of that struct is a subtype of AbstractGrowthRateMethod

Define a new method dispatch growth_rate(data, time_col, ::NameOfNewMethodType)

Document that method in the growth rate documentation (this page)

If you are unsure how to do any of these steps, feel free to open an issue on GitHub asking for a new growth rate method and explaining how the method should work.

1Zwietering MH, Jongenburger I, Rombouts FM, van't Riet K. Modeling of the bacterial growth curve. Appl Environ Microbiol. 1990;56(6):1875-81. doi: 10.1128/aem.56.6.1875-1881.1990.

2Kahm M, Hasenbrink G, Lichtenberg-Fraté H, Ludwig J, Kschischo M. grofit: Fitting Biological Growth Curves with R. Journal of Statistical Software. 2010;33(7):1–21. doi: 10.18637/jss.v033.i07.