Other Useful Functions

There are a couple more functions in ESM that may be generally useful for plate reader data. These are the at and between functions (and there sister functions at_time and between_times).

ESM.at — Function

at(df, range_df, target_value)

Return the values of df at the index of range_df where range_df is equal to target_value.

If no value in range_df is equal to target_value, return the last recording before the target_value.

If range_df has one column, the same index will be applied to all columns of df. If range_df has multiple columns, the column names should match those of df and a new index will be computed for each column.

Arguments:

df::DataFrame: The DataFrame you are getting the indexes from.

range_df: The data frame by which to filter.

target_value::Number: Defines the threshold value for filtering.

ESM.between — Function

between(df, range_df; min_value=-Inf, max_value=Inf)
between(df; min_value=-Inf, max_value=Inf)

Replace out of range values with missing.

Out of range values are determined by min_value .<= range_col .<= max_value. If range_col is not provided, by min_value .<= df .<= max_value.

If range_col has one column, the associate row indexes with min_value .<= range_col .<= max_value are broadcast along all columns of df. If range_col has multiple columns, the column names should match those of df.

Arguments:

df::DataFrame: DataFrame to subset.

range_df::DataFrame: DataFrame of values to filter by.

min_value=-Inf: Minimum value.

max_value=Inf: Maximum value.

ESM.at_time — Function

at_time(df, time_col, time_point)

Returns values at a specific time point. If no value at specific timepoint, return the last recording before the timepoint. If time_point < minimum(time_col), return nothing.

Arguments:

df::DataFrame: DataFrame on which to work.

time_col::DataFrame: Time values in a DataFrame.

time_point::Float64: Time point in mins at which to report the measurement.

ESM.between_times — Function

between_times(df, time_col; mint=-Inf, maxt=Inf)

Returns the DataFrame between two timepoints.

All other data is set to missing.

Arguments:

df::DataFrame: DataFrame to work on.

time_col::DataFrame: DataFrame of time values.

mint::Float64=-Inf: Minimum time in mins.

maxt::Float64=Inf: Max time in mins.

They allow you to remove elements of your data (say removing low OD values to avoid returning a growth rate at low OD when the data is very noisy).

We have a dataset contaning fluorescence (flu and flu_times) and optical density (od and od_times).

julia> flu97×96 DataFrame
 Row │ a1        a2        a3        a4        a5        a6        a7        a ⋯
     │ Float64?  Float64?  Float64?  Float64?  Float64?  Float64?  Float64?  F ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │    852.0     958.0     924.0     989.0     876.0     963.0     994.0    ⋯
   2 │    887.0     941.0     875.0     896.0     894.0     980.0     975.0
   3 │    873.0     913.0     874.0     975.0     915.0     963.0     930.0
   4 │    867.0     928.0     875.0     913.0     900.0     944.0     884.0
   5 │    851.0     874.0     922.0     907.0     859.0     835.0     849.0    ⋯
   6 │    889.0     965.0     868.0     920.0     868.0     852.0     851.0
   7 │    895.0     937.0     936.0     905.0     842.0     818.0     859.0
   8 │    871.0     926.0     903.0     891.0     914.0     817.0     776.0
  ⋮  │    ⋮         ⋮         ⋮         ⋮         ⋮         ⋮         ⋮        ⋱
  91 │    840.0    5563.0    5510.0    5639.0    5521.0    5570.0    5608.0    ⋯
  92 │    852.0    5611.0    5450.0    5528.0    5521.0    5516.0    5741.0
  93 │    839.0    5775.0    5400.0    5543.0    5459.0    5574.0    5684.0
  94 │    860.0    5699.0    5596.0    5706.0    5573.0    5489.0    5709.0
  95 │    885.0    5761.0    5433.0    5602.0    5577.0    5593.0    5768.0    ⋯
  96 │    837.0    5881.0    5656.0    5616.0    5492.0    5535.0    5675.0
  97 │    895.0    5723.0    5654.0    5679.0    5559.0    5664.0    5795.0
                                                  89 columns and 82 rows omitted
julia> flu_times97×1 DataFrame
 Row │ time          
     │ Float64?      
─────┼───────────────
   1 │ 544000.0
   2 │      1.144e6
   3 │      1.744e6
   4 │      2.344e6
   5 │      2.944e6
   6 │      3.544e6
   7 │      4.144e6
   8 │      4.744e6
  ⋮  │       ⋮
  91 │      5.4544e7
  92 │      5.5144e7
  93 │      5.5744e7
  94 │      5.6344e7
  95 │      5.6944e7
  96 │      5.7544e7
  97 │      5.8144e7
      82 rows omitted
julia> od97×96 DataFrame
 Row │ a1        a2        a3        a4        a5        a6        a7        a ⋯
     │ Float64?  Float64?  Float64?  Float64?  Float64?  Float64?  Float64?  F ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │    0.134     0.125     0.124     0.123     0.135     0.128     0.118    ⋯
   2 │    0.133     0.125     0.123     0.122     0.135     0.13      0.118
   3 │    0.133     0.125     0.123     0.131     0.141     0.144     0.128
   4 │    0.133     0.127     0.126     0.137     0.147     0.152     0.138
   5 │    0.133     0.128     0.127     0.14      0.149     0.155     0.141    ⋯
   6 │    0.133     0.129     0.128     0.141     0.15      0.156     0.143
   7 │    0.133     0.13      0.13      0.142     0.151     0.159     0.143
   8 │    0.133     0.13      0.129     0.143     0.151     0.158     0.142
  ⋮  │    ⋮         ⋮         ⋮         ⋮         ⋮         ⋮         ⋮        ⋱
  91 │    0.131     0.846     0.853     0.853     0.86      0.864     0.854    ⋯
  92 │    0.131     0.844     0.851     0.85      0.857     0.861     0.851
  93 │    0.131     0.843     0.849     0.847     0.854     0.858     0.848
  94 │    0.131     0.841     0.846     0.845     0.853     0.857     0.847
  95 │    0.131     0.838     0.844     0.844     0.85      0.855     0.844    ⋯
  96 │    0.131     0.837     0.843     0.842     0.849     0.851     0.843
  97 │    0.131     0.837     0.841     0.839     0.845     0.85      0.84
                                                  89 columns and 82 rows omitted
julia> od_times97×1 DataFrame
 Row │ time          
     │ Float64?      
─────┼───────────────
   1 │ 494000.0
   2 │      1.094e6
   3 │      1.694e6
   4 │      2.294e6
   5 │      2.894e6
   6 │      3.494e6
   7 │      4.094e6
   8 │      4.694e6
  ⋮  │       ⋮
  91 │      5.4494e7
  92 │      5.5094e7
  93 │      5.5694e7
  94 │      5.6294e7
  95 │      5.6894e7
  96 │      5.7494e7
  97 │      5.8094e7
      82 rows omitted

The at_time and between_times functions

The first thing we will look at is how to index the data by time. Below, we will access the od data at 5 minutes and then all of the od data between 3 and 24 minutes.

julia> at_time(od, od_times, 5)┌ Warning: No values found at or before 5.
└ @ ESM ~/work/esm/esm/src/main.jl:231
0×96 DataFrame
 Row │ a1        a2        a3        a4        a5        a6        a7        a ⋯
     │ Float64?  Float64?  Float64?  Float64?  Float64?  Float64?  Float64?  F ⋯
─────┴──────────────────────────────────────────────────────────────────────────
                                                              89 columns omitted
julia> between_times(od, od_times; min_value = 3, max_value = 24)ERROR: MethodError: no method matching between_times(::DataFrames.DataFrame, ::DataFrames.DataFrame; min_value::Int64, max_value::Int64)
This method does not support all of the given keyword arguments (and may not support any).

Closest candidates are:
  between_times(::DataFrames.DataFrame, ::DataFrames.DataFrame; mint, maxt) got unsupported keyword arguments "min_value", "max_value"
   @ ESM ~/work/esm/esm/src/main.jl:146

For at_time, we returned a single row data frame with the od data at 5 minutes (techinically the last recording before 5 minutes).

For between_times, all data that was recorded before 3 minutes and after 24 minutes was set to missing. This means it will be ignored by most functions in ESM.

The at function

Next, we will look at the more general at function. We will use it to get the time when the OD is 0.5 for well A2.

julia> at(od_times, od[:, [:a2]], 0.5)1×1 DataFrame
 Row │ time     
     │ Float64  
─────┼──────────
   1 │ 2.0894e7

We can also use it to get the OD for all wells, when the OD for well A2 is 0.5.

julia> at(od, od[:, [:a2]], 0.5)1×96 DataFrame
 Row │ a1       a10      a11      a12      a2       a3       a4       a5       ⋯
     │ Float64  Float64  Float64  Float64  Float64  Float64  Float64  Float64  ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │   0.132    0.493    0.541    0.668    0.497    0.444    0.443    0.449  ⋯
                                                              88 columns omitted

Finally, we could use it to get the fluorescence for each well when the OD of that well is 0.5. Note that some wells (like A1) are blanks and won't reach 0.5, so the last recording will be used.

julia> at(flu, od, 0.5)1×96 DataFrame
 Row │ a1       a10      a11      a12      a2       a3       a4       a5       ⋯
     │ Float64  Float64  Float64  Float64  Float64  Float64  Float64  Float64  ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │   895.0   1602.0   1445.0   1346.0   2256.0   2060.0   2096.0   2079.0  ⋯
                                                              88 columns omitted

This final version requires a few assumptions. It requires that the column names of the two data frames (od and flu) match. It also assumes that the columns are recorded at roughly the same time.

Thinking about monotonicity

Since we are finding the last row index where OD≤0.5, if OD decreases later on in the experiment and drops below 0.5, we might return that row instead. It's common for OD data to be non-monotonic (both increasing and decreasing at different times). If this is a problem, it may be worth first using between_times to reduce the data to a monotonic subset.

The between function

Finally, we will look at the general between function. It provides a similar interface to the at function, returning a data frame where values outside the range are set to missing.

julia> between(od_times, od[:, [:a2]]; min_value=0.125, max_value=0.8)97×1 DataFrame
 Row │ time          
     │ Float64?      
─────┼───────────────
   1 │  494000.0
   2 │       1.094e6
   3 │       1.694e6
   4 │       2.294e6
   5 │       2.894e6
   6 │       3.494e6
   7 │       4.094e6
   8 │       4.694e6
  ⋮  │       ⋮
  91 │ missing       
  92 │ missing       
  93 │ missing       
  94 │ missing       
  95 │ missing       
  96 │ missing       
  97 │ missing       
      82 rows omitted
julia> between(od, od[:, [:a2]]; min_value=0.125, max_value=0.8)97×96 DataFrame
 Row │ a1           a2           a3           a4           a5           a6     ⋯
     │ Float64?     Float64?     Float64?     Float64?     Float64?     Float6 ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │       0.134        0.125        0.124        0.123        0.135         ⋯
   2 │       0.133        0.125        0.123        0.122        0.135
   3 │       0.133        0.125        0.123        0.131        0.141
   4 │       0.133        0.127        0.126        0.137        0.147
   5 │       0.133        0.128        0.127        0.14         0.149         ⋯
   6 │       0.133        0.129        0.128        0.141        0.15
   7 │       0.133        0.13         0.13         0.142        0.151
   8 │       0.133        0.13         0.129        0.143        0.151
  ⋮  │      ⋮            ⋮            ⋮            ⋮            ⋮            ⋮ ⋱
  91 │ missing      missing      missing      missing      missing      missin ⋯
  92 │ missing      missing      missing      missing      missing      missin
  93 │ missing      missing      missing      missing      missing      missin
  94 │ missing      missing      missing      missing      missing      missin
  95 │ missing      missing      missing      missing      missing      missin ⋯
  96 │ missing      missing      missing      missing      missing      missin
  97 │ missing      missing      missing      missing      missing      missin
                                                  91 columns and 82 rows omitted
julia> between(flu, od; min_value=0.125, max_value=0.8)97×96 DataFrame
 Row │ a1        a2         a3         a4         a5         a6         a7     ⋯
     │ Float64?  Float64?   Float64?   Float64?   Float64?   Float64?   Float6 ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │    852.0      958.0  missing    missing        876.0      963.0  missin ⋯
   2 │    887.0      941.0  missing    missing        894.0      980.0  missin
   3 │    873.0      913.0  missing        975.0      915.0      963.0      93
   4 │    867.0      928.0      875.0      913.0      900.0      944.0      88
   5 │    851.0      874.0      922.0      907.0      859.0      835.0      84 ⋯
   6 │    889.0      965.0      868.0      920.0      868.0      852.0      85
   7 │    895.0      937.0      936.0      905.0      842.0      818.0      85
   8 │    871.0      926.0      903.0      891.0      914.0      817.0      77
  ⋮  │    ⋮          ⋮          ⋮          ⋮          ⋮          ⋮          ⋮  ⋱
  91 │    840.0  missing    missing    missing    missing    missing    missin ⋯
  92 │    852.0  missing    missing    missing    missing    missing    missin
  93 │    839.0  missing    missing    missing    missing    missing    missin
  94 │    860.0  missing    missing    missing    missing    missing    missin
  95 │    885.0  missing    missing    missing    missing    missing    missin ⋯
  96 │    837.0  missing    missing    missing    missing    missing    missin
  97 │    895.0  missing    missing    missing    missing    missing    missin
                                                  90 columns and 82 rows omitted

There is also a simpler way to reduce the OD data to between two thresholds by only providing the od data frame.

julia> between(od; min_value=0.1, max_value=0.8)97×96 DataFrame
 Row │ a1        a2           a3           a4           a5           a6        ⋯
     │ Float64?  Float64?     Float64?     Float64?     Float64?     Float64?  ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │    0.134        0.125        0.124        0.123        0.135        0.1 ⋯
   2 │    0.133        0.125        0.123        0.122        0.135        0.1
   3 │    0.133        0.125        0.123        0.131        0.141        0.1
   4 │    0.133        0.127        0.126        0.137        0.147        0.1
   5 │    0.133        0.128        0.127        0.14         0.149        0.1 ⋯
   6 │    0.133        0.129        0.128        0.141        0.15         0.1
   7 │    0.133        0.13         0.13         0.142        0.151        0.1
   8 │    0.133        0.13         0.129        0.143        0.151        0.1
  ⋮  │    ⋮           ⋮            ⋮            ⋮            ⋮            ⋮    ⋱
  91 │    0.131  missing      missing      missing      missing      missing   ⋯
  92 │    0.131  missing      missing      missing      missing      missing  
  93 │    0.131  missing      missing      missing      missing      missing  
  94 │    0.131  missing      missing      missing      missing      missing  
  95 │    0.131  missing      missing      missing      missing      missing   ⋯
  96 │    0.131  missing      missing      missing      missing      missing  
  97 │    0.131  missing      missing      missing      missing      missing  
                                                  91 columns and 82 rows omitted