API

ESM.at_od — Method

at_od(od_df,target_df,target_od)

A function to return a set of values from a different dataframe based on another. In theory this can be used for many types of data, but this adds the functionality to target a specific OD and return the values from another dataframe at that index.

Args:

od_df=<DataFrame>: The DataFrame you are getting the indexes from.

target_df=<DataFrame>: The DataFrame you want to get the values from.

target_od=<Float64>: The value at which you are limiting the data to

ESM.at_time — Method

at_time(df, time_col, time_point)

Returns values at a specific time point. If no value at specific timepoint, return the last recording before the timepoint. If timepoint < minimum(timecol), return nothing.

Args:

df=<DataFrame>: DataFrame on which to work.

time_col=<DataFrame>: Time values in a DataFrame.

time_point=<Float64>: Time point in mins at which to report the measurement.

ESM.between_times — Method

between_times(df,time_col;mint,maxt)

Returns the DataFrame between two timepoints.

Args:

df=<DataFrame>: DataFrame to work on.

time_col=<DataFrame>: DataFrame of time values.

mint=<Float64>: Minimum time in mins.

maxt=<Float64>: Max time in mins.

ESM.comonicon_install — Method

comonicon_install(;kwargs...)

Install the CLI manually. This will use the default configuration in Comonicon.toml, if it exists. For more detailed reference, please refer to Comonicon documentation.

ESM.comonicon_install_path — Method

comonicon_install_path(;[yes=false])

Install the PATH and FPATH to your shell configuration file. You can use comonicon_install_path(;yes=true) to skip interactive prompt. For more detailed reference, please refer to Comonicon documentation.

ESM.density_gate — Function

density_gate(data,channels=[];gate_frac=0.65,nbins=1024,outside=false)

Function to do density gaing using a KDE.

Args:

data::Dict: Dict returned by RFI.

channels::Vector{String}: Vector of channels to use for gating.

gate_frac::Float64: Amount of samples to keep.

nbins::Int64: Number of histogram bins.

outside::Bool: Whether to return the outside data (only for internal troubleshooting not fore general usage).

ESM.df2time — Method

df2time(time_col)

Converts a time column in a DataFrame with String elements to time in seconds.

Args:

time_col=<DataFrame>: DataFrame with time values.

ESM.doubling_time — Method

doubling_time(df, time_col;max_od)

Calculates the growth rate/doubling time of a given dataframe. returns in min^-1

equation = (tend - tstart)/log2(maxod/min_od)

Args:

df=<DataFrame>: DataFrame containing the data.

time_col=<DatFrame>: DataFrame containing the times.

max_od=<Float64>: Maximum OD. Defaults to 0.4 - min_od is 1/4 of this value.

ESM.expand_group — Method

expand_group(group::AbstractString)

Expand a condensed group name into the sample IDs it contains. For example, if the group is "plate0[1,2][a:c]2", it will return all sample IDs from plate 1 to 2 and wells a2, b2, and c2.

Args:

group::AbstractString: The group name to expand.

Returns:

Vector{String}: A vector of sample IDs contained in the group.

ESM.expand_groups — Method

expand_groups(groups::AbstractString)

ESM.extract_flow — Method

extract_flow(fcs,chan)

Extract the metadata from a fcs file.

Args:

fcs::FCSFile: The read FCS file

chan::String: The channel being read from

ESM.filter_col — Method

filter_col(df, reg_l)

Filters a df by a set of columns or parts of columns given in regl. regl is joined to form regex using |.

Args:

df=<DataFrame> DataFrame to filter.

reg_l=<Array{String}>: Array of strings to filter by.

ESM.filter_row — Method

filter_row(es,group)

Filter the original es.samples dataframe by a specific group defined in the esm.

Args:

es=<esm_zones>: esm zones data type. global

group=<string || vector>: what group/groups to filter by.

ESM.find_group — Method

find_group(es,grn)

Finds a specific group from the original es.groups dataframe and returns the sample names.

Args:

es=<esm_zones>: the original es object.

grn=<String>: A string of a group name which can be used to filter the original dataframe.

ESM.form_df — Method

form_df(df)

Creates the data frame from the dataframe it is passed. This is to turn the es.samples.values arrays into dfs.

Args:

df=<DataFrame>: A dataframe of es.samples to parse.

ESM.growth_rate — Method

growth_rate(df; window_size=10)

Calculates the growth rate of a given dataframe. Returns in min^-1 using base e.

Args:

df=<DataFrame>: DataFrame containing the data.

window_size=10: Size of the window (in minutes) to use for calculating the growth rate. Defaults to 10.

ESM.high_low — Method

high_low(data;chans=[],maxr=missing,minr=missing)

Gate hig/low the data.

Args:

data::Dict: Dictionary of data given by to_rfi

chans::Vector: Channels to work over.

maxr::Int64: max range

minr::Int64: min range

ESM.index_between_vals — Method

index_between_vals(df; minv, maxv)

Returns the indexes between two points so that these can be later isolated. This is in dict form to allow for individual points to be separated out. Both ends are inclusive.

If minv is larger than all values in the column, or maxv smaller, will return (nothing,nothing). Args:

df=<DataFrame>: DataFrame to work on.

minv=<Float64>: Minimum value.

maxv=<Float64>: Maximum value.

ESM.process_fcs — Method

process_fcs(group::String,gate_channels::Vector,out_channels::Vector{String};gate_frac=0.65,nbins=1024,hl_channels=[],rfi=true,dense=true,hl=true,maxr=missing,minr=missing)

Processes fcs files from a given group. Can just calculate RFI, but also can do high/low gating and density gating.

Args:

group::String: defines the group to work from.

gate_channels::Vector: defines the channels to work from.

out_channels::Vector{String}: defines the output channel.

gate_frac=0.65: defines the fraction of population to keep.

nbins=1024: defines the number of bins to use - mostly redundant.

hl_channels=[]: which channels to use for high/low gating.

rfi=true: whether to do RFI.

dense=true: whether to do density gating.

hl=true: whether to do high low gating.

maxr=missing: max val for high/low gating.

minr=missing: min val for high/low gating.

ESM.produce_views — Method

produce_views(es, trans_meta_map;to_out=[])

Produces the views specified in to_out else all views in esm.views.

Args:

es=<esm_zones>: The esm_zones object (global)

trans_meta_map=<Dict>: A dictionary mapping transformations to their names.

to_out=<Array>: A list of all the views to be produced.

ESM.read_data — Method

read_data(filen)

Read the data from path filen into the ESM structure for writing.

Args:

filen::String: String to where the excel file is.

ESM.read_esm — Method

read_esm(filen)

Reads and processes the esm file produced at a file name and creates the esm_zones struct.

Args:

filen=<String>

esm.samples DataFrame cols: |name | channel | type | values | meta | group... |

name - name.channel

channel - data channel

type - type of sample - time series or population

values - Arrays of data values.

meta - any meta data associated with the data.

group - boolean columns associating the groups with samples.

ESM.read_flow — Method

read_flow(samples, sample_dict,channels,broad_g,channel_map)

Read the flow cytometry data from a file.

Args:

samples::DataFrame: DataFrame of samples to read from.

sample_dict::Dict: Output sample Dict.

channels::Vector: Channels to read.

broad_g::Vector: The physical group that this is part of.

channel_map::Dict: The names to replace the old channel names with.

ESM.read_multipr_file — Method

read_multipr_file(filen,ptype,channels,channel_map)

Function for reading files containing multiple reads from multple channels from a single CSV file. Returns a Dictionary of DataFrames with keys being the channels. Currently works for Agilent (agilent), Tecan (tecan) and Spectramax (spectramax). Presently uses regexes to seperate the files.

Args:

filen::String: File to read from.

ptype::String: The plate reader type being used.

channels::Vector: Channels to be read in.

channel_map: The new names for the channels.

ESM.read_pr — Method

read_pr(samples,sample_dict,channels,broad_g,channel_map)

Reads plate reader data - the master controller function.

Args:

samples::DataFrame: Dataframe of samples to be read into sample_dict.

sample_dict::Dict: A dictionary of the samples.

channels::Vector: Channels to use.

broad_g::Vector: The larger plate that this belongs to.

channel_map::Dict: What to rename the channel to.

ESM.read_sep_chans_pr — Method

read_sep_chans_pr(channel_map,loc,channels)

Function to read a separated set of csvs into the ESM. All files are found in one dir and the names of them are the channel names.

Args:

channel_map::Dict: What to rename the channel to once read.

loc::String: Where the file actually is.

channels::Vector: The channels to use to read the files.

ESM.remove_subcols — Method

remove_subcols(df, sub)

Gets rid of certain sub col ids in the col names.

e.g.

remove flo
a1.flo -> a1

ESM.sexp_to_nested_list — Method

sexp_to_nested_list(sexp,es,trans_meta_map)

Recursively converts the parsed equations to julia code and produces the correct dataframes over which to operate. This also calls any unprocessed transformations.

Args:

sexp - Expression or part of expression to be decomposed.

es - The esm_zones data type that contains the data (global variable).

transmetamap - The transformation map of the parsed transformations.

ESM.summarise — Function

esm summarise

Summarise a data file (.esm, plate reader, .fcs, etc.).

Options

-f, --file=<String>: The data file to be summarised.

-t, --type=<String>: The type of data file. Options are "auto" (default), "esm", "spectramax", "biotek", "fcs". If "auto" is selected, the type will be inferred from the file extension (or raise an error if not possible).

Flags

-p, --plot: Produce plots of the data. Only available for some types.

ESM.summarise_biotek — Method

summarise_biotek(file)

Summarises a Biotek output file.

Args:

file::String: Path to the Biotek file.

plot::Bool: Whether to plot the data.

ESM.summarise_esm — Method

summarise_esm(file)

Summarises the contents of an esm file.

Args:

file::String: The path to the esm file to be summarised.

plot::Bool: Whether to plot the timeseries data.

ESM.summarise_spectramax — Method

summarise_spectramax(file)

Summarises a SpectraMax output file.

Args:

file::String: Path to the SpectraMax file.

plot::Bool: Whether to plot the data.

ESM.template — Function

esm template

Produce a template excel file for data entry into the ESM.

Options

-o, --output-path=<String>: The path to create the template in. Defaults to ESM.xlsx in the current directory.

ESM.to_rfi — Method

to_rfi(sample_name;chans=[])

Calculates relative fluorescence of given sample.

Args:

sample_name::String: channel to use.

chans::Vector: vector of channels to keep.

ESM.translate — Function

esm translate

Translates the completed .xlsx template file to a .esm file.

Options

-e, --excel=<String>: The .xlsx template file to be read.

-t, --target=<String>: The name of the output .esm file.

ESM.view_to_csv — Method

view_to_csv(es,trans_meta_map;out_dir,to_out)

Controls the calling of the view producer and writes the outputted views to a csv file.

Args:

es=<esm_zones>: The original esm struct

trans_meta_map=<Dict>: The transformations that have been parsed at the top level.

out_dir=<String>: The specified output dir - defaults to nothing.

to_out=<Array{String}>: The views to be processed.

ESM.views — Function

esm views

Produce and save the views from a .esm file.

Options

-e, --esm-file=<String>: The .esm file to be read.

-v, --view=<String>: The view to be produced. If not specified, all views will be produced.

-o, --output-dir=<String>: The directory to save the output(s) to. Defaults to the current directory.

ESM.write_esm — Method

write_esm(esm_dict, name)

Write the esmdict from readdata() to the path/file name.

Args:

esm_dict::Dict: Dictionary describing the esm.

name::String: Path to deposit the ESM to.