API

Base.read — Method

read(file::AbstractString, ::AbstractPlateReader; channels)

Read data from plate reader file file of assuming the format of AbstractPlateReader.

Arguments:

file::AbstractString: File to read from.

::AbstractPlateReader: Plate reader type.

channels: Vector of channels (as strings) to be read in. Defaults to nothing (all channels).

Base.summary — Method

summary(file, ptype::AbstractESMDataType; plot=false)

Summarise a data file (.esm, plate reader, .fcs, etc.).

Arguments

file::AbstractString: The data file to be summarised.

ptype::AbstractESMDataType: The type of data file.

plot::Bool=false: Produce plots of the data. Defaults to false.

ESM.at_od — Method

at_od(od_df,target_df,target_od)

A function to return a set of values from a different dataframe based on another. In theory this can be used for many types of data, but this adds the functionality to target a specific OD and return the values from another dataframe at that index.

Arguments:

od_df::DataFrame: The DataFrame you are getting the indexes from.

target_df::DataFrame: The DataFrame you want to get the values from.

target_od::Number: The value at which you are limiting the data to

ESM.at_time — Method

at_time(df, time_col, time_point)

Returns values at a specific time point. If no value at specific timepoint, return the last recording before the timepoint. If timepoint < minimum(timecol), return nothing.

Arguments:

df::DataFrame: DataFrame on which to work.

time_col::DataFrame: Time values in a DataFrame.

time_point::Float64: Time point in mins at which to report the measurement.

ESM.between_times — Method

between_times(df,time_col;mint=-Inf,maxt=Inf)

Returns the DataFrame between two timepoints.

Arguments:

df::DataFrame: DataFrame to work on.

time_col::DataFrame: DataFrame of time values.

mint::Float64=-Inf: Minimum time in mins.

maxt::Float64=Inf: Max time in mins.

ESM.calibrate — Method

calibrate(data, time_col, method::AbstractCalibrationMethod)

Calibrate data, for example, to remove background OD signal.

Arguments:

data::DataFrame: DataFrame containing the data to be calibrated.

time_col::Union{DataFrame, Nothing}: Time column associated with the data.

method::AbstractCalibrationMethod: Method to use for calibration.

ESM.comonicon_install — Method

comonicon_install(;kwargs...)

Install the CLI manually. This will use the default configuration in Comonicon.toml, if it exists. For more detailed reference, please refer to Comonicon documentation.

ESM.comonicon_install_path — Method

comonicon_install_path(;[yes=false])

Install the PATH and FPATH to your shell configuration file. You can use comonicon_install_path(;yes=true) to skip interactive prompt. For more detailed reference, please refer to Comonicon documentation.

ESM.df2time — Method

df2time(time_col)

Converts a time column in a DataFrame with String elements to time in seconds.

Arguments:

time_col::DataFrame: DataFrame with time values.

ESM.doubling_time — Method

doubling_time(df, time_col, method::AbstractGrowthRateMethod; kwargs...)

Calculates the doubling time of a given DataFrame. Returns result in minutes.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod: Method to use for calculating growth rate.

kwargs...: Additional keyword arguments to pass to the growth rate method.

ESM.event_count — Method

event_count(data)

Count the number of events in the flow cytometry data.

Arguments:

data::Dict: Dict returned by to_rfi.

ESM.expand_group — Method

expand_group(group::AbstractString)

Expand a condensed group name into the sample IDs it contains. For example, if the group is "plate0[1,2][a:c]2", it will return all sample IDs from plate 1 to 2 and wells a2, b2, and c2.

Arguments:

group::AbstractString: The group name to expand.

Returns:

Vector{String}: A vector of sample IDs contained in the group.

ESM.expand_groups — Method

expand_groups(groups::AbstractString)

Iteratively call expand_group on a comma-separated list of group names.

ESM.extract_flow — Method

extract_flow(fcs,chan)

Extract the metadata from a FCSFiles.FlowSample.

Arguments:

fcs: The parsed FCS file.

chan::String: The channel being read from.

ESM.filter_col — Method

filter_col(df, reg_l)

Filters a df by a set of columns or parts of columns given in regl. regl is joined to form regex using |.

Arguments:

df::DataFrame: DataFrame to filter.

reg_l::Vector{String}: Vector of strings to filter by.

ESM.filter_row — Method

filter_row(es,group)

Filter the original es.samples dataframe by a specific group defined in the esm.

Arguments:

es::esm_zones: esm zones data type.

group::Union{String,Vector}: what group/groups to filter by.

ESM.find_group — Method

find_group(es,grn)

Finds a specific group from the original es.groups dataframe. Returns the sample names.

Arguments:

es::esm_zones: The data set to search.

grn::String: A string of a group name which can be used to filter the original dataframe.

ESM.fit_ellipse — Method

fit_ellipse(x, y) → a, b, θ, x₀, y₀, p

Fit an ellipse into the 2D data defined by (x, y) using ordinary least squares.

Return: semi-major axis length, semi-minor axis length, ellipse rotation, center coordinates and a parameter container for quadratic form of ellipse. The first five parameters represent the parametric form of an ellipse, which may have an arbitrary rotation and translation w.r.t. the origin:

x(t) = cos(θ)*a*cos(t) - sin(θ)*b*sin(t) + x₀
y(t) = sin(θ)*a*cos(t) + cos(θ)*b*sin(t) + y₀
t ∈ [0, 2π)

Use ellipse_from_parametric(a, b, θ, x₀, y₀) → x, y to construct points along the ellipse. Notice that always two possible ellipses can fit the data, one with a, b, θ and one with b, a, θ-π/2.

The quadratic form of the ellipse is created using p, and is given by the form (x^2, x*y, y^2, x, y) ⋅ p = 1.

Code modified from: https://www.matecdev.com/posts/julia-least-squares-qr.html using a lot of stuff from: https://en.wikipedia.org/wiki/Ellipse#General_ellipse

ESM.form_df — Method

form_df(df)

Creates the data frame from the dataframe it is passed. This is to turn the es.samples.values arrays into dfs.

Arguments:

df::DataFrame: A dataframe of es.samples to parse.

ESM.gate — Function

gate(data, method::AbstractGatingMethod)

Filter data to only include events within the gate defined by method.

Arguments:

data::Dict: Dict returned by to_rfi.

method::AbstractGatingMethod: The method and settings to use for gating.

ESM.gated_proportion — Method

gated_proportion(data, gate)
gated_proportion(data_before, data_after)

Calculate the proportion of events remaining after gating.

Arguments:

data::Dict: Dict returned by to_rfi.

gate::AbstractGatingMethod: A gating method to report on.

data_before::Dict: Dict returned by to_rfi before gating.

data_after::Dict: Dict returned by to_rfi after gating.

ESM.growth_rate — Method

growth_rate(df, time_col, method::AbstractGrowthRateMethod)

Calculates the growth rate of a given dataframe. Returns in min^-1 using base e.

Arguments:

df::DataFrame: DataFrame containing the data.

time_col::DataFrame: DataFrame containing the times.

method::AbstractGrowthRateMethod: Method to use for calculating growth rate.

ESM.index_between_vals — Method

index_between_vals(df; minv=-Inf, maxv=Inf)

Returns the indexes between two points so that these can be later isolated. This is in dict form to allow for individual points to be separated out. Both ends are inclusive.

If minv is larger than all values in the column, or maxv smaller, will return (nothing,nothing).

Arguments:

df::DataFrame: DataFrame to work on.

minv::Float64=-Inf: Minimum value.

maxv::Float64=Inf: Maximum value.

ESM.produce_views — Method

produce_views(es, trans_meta_map;to_out=[])

Produces the subset of views specified in to_out or all views if unspecified.

Arguments:

es::esm_zones: The esm_zones object

trans_meta_map::Dict: A dictionary mapping transformations to their names.

to_out::Vector{String}: A list of all the views to be produced.

ESM.read_data — Method

read_data(file::AbstractString)

Read the data from path file into the correct structure.

ESM.read_esm — Method

read_esm(file::AbstractString)

Parse an esm file found at file into an esm_zones object.

ESM.read_flow — Method

read_flow(samples, sample_dict,channels,broad_g,channel_map)

Read the flow cytometry data from a file.

Arguments:

samples::DataFrame: DataFrame of samples to read from.

sample_dict::Dict: Output sample Dict.

channels::Vector: Channels to read.

broad_g::Vector: The physical group that this is part of.

channel_map::Dict: The names to replace the old channel names with.

ESM.read_multipr_file — Method

read_multipr_file(file,ptype,channels,channel_map)

Function for reading files containing multiple reads from multple channels from a single CSV file. Returns a Dictionary of DataFrames with keys being the channels.

Arguments:

file::String: File to read from.

ptype::String: The plate reader type being used.

channels::Vector: Channels to be read in.

channel_map: The new names for the channels.

ESM.read_pr — Method

read_pr(samples,sample_dict,channels,broad_g,channel_map)

Reads plate reader data.

Arguments:

samples::DataFrame: Dataframe of samples to be read into sample_dict.

sample_dict::Dict: A dictionary of the samples.

channels::Vector: Channels to use.

broad_g::Vector: The larger plate that this belongs to.

channel_map::Dict: What to rename the channel to.

ESM.remove_subcols — Method

remove_subcols(df, sub)

Gets rid of certain sub col ids in the col names.

e.g.

remove flo
a1.flo -> a1

ESM.sexp_to_nested_list — Method

sexp_to_nested_list(sexp,es,trans_meta_map)

Recursively converts the parsed equations to julia code and produces the correct dataframes over which to operate. This also calls any unprocessed transformations.

Arguments:

sexp: Expression or part of expression to be decomposed.

es: The esm_zones data type that contains the data.

transmetamap: The transformation map of the parsed transformations.

ESM.summarise — Function

esm summarise

Summarise a data file (.esm, plate reader, .fcs, etc.).

Options

-f, --file=<String>: The data file to be summarised.

-t, --type=<String>: The type of data file. Options are "auto" (default), "esm", "spectramax", "biotek", "generic", "fcs". If "auto" is selected, the type will be inferred from the file extension (or raise an error if not possible).

Flags

-p, --plot: Produce plots of the data. Only available for some types.

ESM.template — Function

esm template

Produce a template excel file for data entry into the ESM.

Options

-o, --output-path=<String>: The path to create the template in. Defaults to ESM.xlsx in the current directory.

ESM.to_rfi — Method

to_rfi(sample_name;chans=[])

Calculates relative fluorescence of given sample.

Arguments:

sample_name::String: channel to use.

chans::Vector: vector of channels to keep.

ESM.translate — Function

esm translate

Translates the completed .xlsx template file to a .esm file.

Options

-e, --excel=<String>: The .xlsx template file to be read.

-t, --target=<String>: The name of the output .esm file.

ESM.view_to_csv — Method

view_to_csv(es,trans_meta_map;out_dir,to_out)

Produces views and writes them to CSV files.

Arguments:

es::esm_zones: The original esm struct

trans_meta_map::Dict: The transformations that have been parsed at the top level.

outdir::String: The specified output dir - defaults to nothing.

to_out::Vector{String}: The views to be processed.

ESM.views — Function

esm views

Produce and save the views from a .esm file.

Options

-e, --esm-file=<String>: The .esm file to be read.

-v, --view=<String>: The view to be produced. All views if not specified.

-o, --output-dir=<String>: The directory to save the output(s) to. Defaults to the current directory.

ESM.write_esm — Method

write_esm(data, file::AbstractString)

Write the esm data to the path file.